Allow setting hostname if index not changed
Allow for the hostname to be set to the default hostname if SRE Tooling
doesn't change the index. Putting a `set -e` at the top of the file will cause
the script to exit immediately after the sre-tooling command is run if the index
didn't need updating. This means that for instances that are correctly indexed
but who's hostname isn't set correctly, the hostname will not be updated.